LayoutTests/fast/canvas/canvas-stroke-zeroSizeGradient-expected.txt

Series of tests to ensure that stroke() paints nothing on canvas when the strokeStyle is set to a zero-size gradient.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS imgdata[0] is 0
PASS imgdata[1] is 255
PASS imgdata[2] is 0
PASS successfullyParsed is true

TEST COMPLETE
